+++ CHANGELOG.md
@@ -8,8 +8,9 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 
 - [AL-3170] Fixed a memory leak issue in ALKConversationViewController.
 
-### Enhancements
 
+### Enhancements
+- [AL-3159] Separated tableview from conversationlist controller into a separate controller.
 - [AL-2856] Dismiss Typing indicator in 30 seconds if stop flag not received
 
 1.3.0